requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.3,<26.0
djangorestframework-simplejwt>=5.4,<6.0
drf-spectacular>=0.27,<1.0
psycopg[binary]>=3.2,<4.0
gunicorn>=22.0,<24.0
django-cors-headers>=4.5,<5.0
